Change to use models facade
