+++ setup.cfg
@@ -1,5 +1,5 @@
 [tool:pytest]
-addopts = -v --cov=open_alchemy --cov=examples/app --cov-config=tests/.coveragerc --flake8 --strict -m "column"
+addopts = -v --cov=open_alchemy --cov=examples/app --cov-config=tests/.coveragerc --flake8 --strict
 flakes-ignore =
     *__init__.py ALL
 markers =
