+++ tests/open_alchemy/conftest.py
@@ -172,20 +172,6 @@ def cleanup_models():
 
 
 @pytest.fixture
-def mocked_models(monkeypatch):
-    """Monkeypatch open_alchemy.models"""
-    mock_models = mock.MagicMock()
-    monkeypatch.setattr(open_alchemy, "models", mock_models)
-    yield mock_models
-
-
-@pytest.fixture
-def _mocked_models(mocked_models):
-    """Suppress unused argument error."""
-    yield mocked_models
-
-
-@pytest.fixture
 def mocked_facades_models(monkeypatch):
     """Monkeypatch open_alchemy.facades.models."""
     mock_models = mock.MagicMock()
